Create pollutantmean.R
Initial solution to assignment.  This solution is a function that first creates a character vector of the files in the directory (returning full names).  This character vector simplifies and automates the process of reading data from the specified monitors (id) since each of the elements of the vector correspond to the appropriate file (monitor id).  Since each of the files maintain the same table structure, a data frame is initialized to store data for each of the specified monitors in the id argument.  The "for" control structure is used with an index variable to iterate through each of the specified id's, reading the data and then row-binding the data to the data frame.  A single data frame is created and the return value of function is calculated as the mean of the pollutant (column in the data frame based on specified argument "pollutant").