Create alt_pollutantmean.R
Alternate solution to assignment, more simple and elegant than the initial solution.  This solution is a function that first creates a character vector of the files in the directory (returning full names).  This character vector simplifies and automates the process of reading data from the specified monitors (id) since each of the elements of the vector correspond to the appropriate file (monitor id).  This solution assumes that the table structure for each of the monitors may not be the same (i.e. the order of the columns can be different, or may have additional or less columns than others).  Therefore, rather than row-binding data into a data frame, only the specified column for the pollutant is stored by first initializing a vector (pollutantdata) and then concatenating relevant pollutant data from the specified id's.  The "for" control structure is used with an index variable to iterate through each of the specified id's, reading the data and then concatenating the data to the pollutantdata vector.  A single vector is created (pollutantdata) and the return value of function is calculated as the mean of the pollutantdata vector.