Create pollutant_corr_complete_cases.R
This function calculates the number of complete cases for the monitor data for each monitor (cases where neither sulfate nor nitrate are NAs) and if the number of complete cases is greater than the threshold argument (threshold can be set in argument, set to 0 as default), calculates the correlation between each pollutant (sulfate and nitrate).  The function returns a vector of correlations for the monitors where the number of complete cases is greater than the threshold.  (Note: Part 2 of the assignment was to create a function to calculate the number of complete cases for each monitor ID.  Rather than source the previous function, this function is constructed to test the condition of good cases within the function without calling the function in Part 2.  An alternative solution is provided that sources and calls the  "complete" function from Part 2.)