Create alt_pollutant_corr_complete_cases.R
Alternative solution that calls the complete function created in Part 2 (complete_nobs.R, which creates a data frame showing each monitor id and the number of complete cases).  This function calculates the number of complete cases for the monitor data for each monitor (cases where neither sulfate nor nitrate are NAs) and if the number of complete cases is greater than the threshold argument (threshold can be set in argument, set to 0 as default), calculates the correlation between each pollutant (sulfate and nitrate).  The function returns a vector of correlations for the monitors where the number of complete cases is greater than the threshold.